Switch from explicit euler to leap from integration scheme.
This has several advantages, even if it introduces additional variables in the form of velocity half steps.
The velocity for the viscosity force needs to be calculated for timestep i, whereas the velocity and
position integration now work with the velocities of timesteps i + 1/2. The advantages are a higher
accuracy of order two and the use of a symplectic scheme, which avoids the energy loss (or build-up)
associated with the explicit or implicit euler schemes.