Fix start distribution of particles.
Instead of randomly distributing them, which can easily lead to absurd values
for physical measures like kinetic energy, we use a regular distribution.
We assume the particles are distributed in a sphere packing with a distance of
each other equal to 2 times the smoothing length.
The smoothing length and the number of particles should follow this approximation:
N * V_s ~= rho_p * V_d
with
N: the number of particles
V_s = (4/3)*pi*(h/kappa)^3 : The volume of the sphere of influence of one particle
rho_p = 0.74048 : The density of a sphere packing
V_d : The volume of the domain length, which is simply 1 for the standard cube
kappa: An overlap coefficient, which should be close to 1.0
h: The smoothing length