Fix issue with zero values for SOD of kernel.
The problem is that the kernel contains divisions by zero if one of the rx, ry, rz or
r arguments is zero. The mathematical formula to describe the forces has no problem
with this, as the seemingly infinite (actually undefined) values is multiplied with
zero in this case, canceling them out zero.
However the implementation of floating point arithmetic does not work the same way and
would evaluate the expression to infinity instead of zero. Therefore we check the
value of the arguments before calling the function.
Additionally we check for a value very close to zero instead of a direct comparison
with zero because the division by very small numbers can have the same effect as
a division by zero. Why? Because floating point arithmatic is tricksy business.