Add small disturbances to initial positions and velocities.
This serves the purpose of breaking symmetry as early as the first time step. Some
of the equations of the particle forces have unwanted behaviour if the difference
in position or velocity between two particles is exactly zero. During the normal run
this is not much of a problem since the difference is rarely zero. Hpwever during
initialisation this occurs much more frequently.