Fix gradient formulation from canonical to symmetric.
One source (see source Intel) approaches the gradient of sizes with three different formulations;
namely canonical, difference and symmetric. While canonical is the least computationally
expensive, it is not symmetric and could not be used to conserve momentum with the
symmetric force trick. Furthermore, the existing implementation of the gradient was wrong,
apparently.