Improve debug view of positions.
We now draw a border region around the standard cube and draw the borders themselves
to show where the cube is. Additionately we add a seperate parameter for the height
of the view that is rendered. This allows the user to see the particles above the
standard cube, as the upper border is open.