Replace parameter class with YAML implementation AND replace double with float.
Unfortunately this commit cannot be easily seperated into two. The first part is done for a more flexible parameter file and reading out values. The second part is done because doubles are not supported by all libraries we are going to use in this project. The additional precision is also of questionable usefulness as the algorithm do not guarantuee any kind of precision anyway.