Put debug renderer on the heap.
It is not entirely clear how the pixel data in the SDL library is stored. It is possible that if the renderer is declared to be on the stack that images that are too large would result in a stack overflow. Working with a pointer puts the renderer on the heap and hopefully avoids this possible stack overflow.